Rename `peace_rt_logic` crate to `peace_rt`.
