+++ Cargo.toml
@@ -19,7 +19,7 @@ test = false
 peace_cfg = { path = "crate/cfg", version = "0.0.1" }
 peace_data = { path = "crate/data", version = "0.0.1" }
 peace_diff = { path = "crate/diff", version = "0.0.1" }
-peace_rt_logic = { path = "crate/rt_logic", version = "0.0.1" }
+peace_rt = { path = "crate/rt", version = "0.0.1" }
 peace_rt_model = { path = "crate/rt_model", version = "0.0.1" }
 
 [dev-dependencies]
