+++ src/lib.rs
@@ -3,5 +3,5 @@
 pub use peace_cfg as cfg;
 pub use peace_data as data;
 pub use peace_diff as diff;
-pub use peace_rt_logic as rt_logic;
+pub use peace_rt as rt;
 pub use peace_rt_model as rt_model;
