+++ workspace_tests/src/lib.rs
@@ -4,7 +4,7 @@ pub(crate) use vec_copy_full_spec::{VecCopyError, VecCopyFullSpec};
 
 mod data;
 mod diff;
-mod rt_logic;
+mod rt;
 mod rt_model;
 
 mod vec_copy_full_spec;
