+++ workspace_tests/src/rt/job_runner.rs
@@ -1,4 +1,4 @@
-use peace::rt_logic::JobRunner;
+use peace::rt::JobRunner;
 
 #[test]
 fn returns_ok_when_job_is_successful() {
